+++ include/patomic/types/ops.h
@@ -1,6 +1,8 @@
 #ifndef PATOMIC_OPS_H
 #define PATOMIC_OPS_H
 
+#include <patomic/types/transaction.h>
+
 #ifdef __cplusplus
 extern "C" {
 #endif
@@ -116,13 +118,93 @@ typedef void (* patomic_opsig_explicit_void_noarg_t) \
 
 
 /*
+ * TRANSACTION OPSIGS
+ * - in-params
+ * - out-params
+ * - result-out-param
+ * - config
+ *
+ * - (except for flag - its only param is the flag)
+ */
+
+typedef void (* patomic_opsig_transaction_store_t) \
+    (volatile void *obj,                           \
+     const void *desired,                          \
+     patomic_transaction_result_t *result,         \
+     patomic_transaction_config_t config);
+
+typedef void (* patomic_opsig_transaction_load_t) \
+    (const volatile void *obj,                    \
+     void *ret,                                   \
+     patomic_transaction_result_t *result,        \
+     patomic_transaction_config_t config);
+
+typedef void (* patomic_opsig_transaction_exchange_t) \
+    (volatile void *obj,                              \
+     const void *desired,                             \
+     void *ret,                                       \
+     patomic_transaction_result_t *result,            \
+     patomic_transaction_config_t config);
+
+typedef int (* patomic_opsig_transaction_cmpxchg_t) \
+    (volatile void *obj,                            \
+     void *expected,                                \
+     const void *desired,                           \
+     patomic_transaction_result_t *result,          \
+     patomic_transaction_config_t config);
+
+typedef int (* patomic_opsig_transaction_test_t) \
+    (const volatile void *obj,                   \
+     size_t offset,                              \
+     patomic_transaction_result_t *result,       \
+     patomic_transaction_config_t config);
+
+typedef int (* patomic_opsig_transaction_test_modify_t) \
+    (volatile void *obj,                                \
+     size_t offset,                                     \
+     patomic_transaction_result_t *result,              \
+     patomic_transaction_config_t config);
+
+typedef void (* patomic_opsig_transaction_fetch_t) \
+    (volatile void *obj,                           \
+     const void *arg,                              \
+     void *ret,                                    \
+     patomic_transaction_result_t *result,         \
+     patomic_transaction_config_t config);
+
+typedef void (* patomic_opsig_transaction_fetch_noarg_t) \
+    (volatile void *obj,                                 \
+     void *ret,                                          \
+     patomic_transaction_result_t *result,               \
+     patomic_transaction_config_t config);
+
+typedef void (* patomic_opsig_transaction_void_t) \
+    (volatile void *obj,                          \
+     const void *arg,                             \
+     patomic_transaction_result_t *result,        \
+     patomic_transaction_config_t config);
+
+typedef void (* patomic_opsig_transaction_void_noarg_t) \
+    (volatile void *obj,                                \
+     patomic_transaction_result_t *result,              \
+     patomic_transaction_config_t config);
+
+typedef void (* patomic_opsig_transaction_flag_t) \
+    (patomic_transaction_flag_t *);
+
+
+/*
  * OP STRUCTS
  * - implicit: each operation is a function pointer (fp)
  *             without a memory order parameter (implicit)
  * - explicit: each operation is a function pointer (fp)
  *             with an explicit memory order parameter
+ * - transaction: each operation is a function pointer (fp)
+ *                with either a transaction result and config,
+ *                or a flag
  */
 
+#undef PATOMIC_DEFINE_ARITHMETIC_OPS_STRUCT
 #define PATOMIC_DEFINE_ARITHMETIC_OPS_STRUCT(opsig_kind, ops_kind) \
     typedef struct {                                               \
         /* in place */                                             \
@@ -139,6 +221,7 @@ typedef void (* patomic_opsig_explicit_void_noarg_t) \
         patomic_##opsig_kind##_fetch_noarg_t fp_fetch_neg;         \
     } patomic_##ops_kind##_arithmetic_t
 
+#undef PATOMIC_DEFINE_BINARY_OPS_STRUCT
 #define PATOMIC_DEFINE_BINARY_OPS_STRUCT(opsig_kind, ops_kind) \
     typedef struct {                                           \
         /* in place */                                         \
@@ -153,6 +236,7 @@ typedef void (* patomic_opsig_explicit_void_noarg_t) \
         patomic_##opsig_kind##_fetch_noarg_t fp_fetch_not;     \
     } patomic_##ops_kind##_binary_t
 
+#undef PATOMIC_DEFINE_BITWISE_OPS_STRUCT
 #define PATOMIC_DEFINE_BITWISE_OPS_STRUCT(opsig_kind, ops_kind) \
     typedef struct {                                            \
         patomic_##opsig_kind##_test_t fp_test;                  \
@@ -161,6 +245,7 @@ typedef void (* patomic_opsig_explicit_void_noarg_t) \
         patomic_##opsig_kind##_test_modify_t fp_test_reset;     \
     } patomic_##ops_kind##_bitwise_t
 
+#undef PATOMIC_DEFINE_XCHG_OPS_STRUCT
 #define PATOMIC_DEFINE_XCHG_OPS_STRUCT(opsig_kind, ops_kind) \
     typedef struct {                                         \
         patomic_##opsig_kind##_exchange_t fp_exchange;       \
@@ -168,28 +253,63 @@ typedef void (* patomic_opsig_explicit_void_noarg_t) \
         patomic_##opsig_kind##_cmpxchg_t fp_cmpxchg_strong;  \
     } patomic_##ops_kind##_xchg_t
 
-#define PATOMIC_DEFINE_OPS_STRUCT(opsig_kind, ops_kind)         \
+typedef struct {
+    patomic_opsig_transaction_flag_t fp_set;
+    patomic_opsig_transaction_flag_t fp_reset;
+} patomic_ops_transaction_flag_t;
+
+typedef struct {
+    char _todo;
+} patomic_ops_transaction_special_t;
+
+#undef PATOMIC_DEFINE_OPS_STRUCTS
+#define PATOMIC_DEFINE_OPS_STRUCTS(opsig_kind, ops_kind)        \
     PATOMIC_DEFINE_ARITHMETIC_OPS_STRUCT(opsig_kind, ops_kind); \
     PATOMIC_DEFINE_BINARY_OPS_STRUCT(opsig_kind, ops_kind);     \
     PATOMIC_DEFINE_BITWISE_OPS_STRUCT(opsig_kind, ops_kind);    \
-    PATOMIC_DEFINE_XCHG_OPS_STRUCT(opsig_kind, ops_kind);       \
-    typedef struct {                                            \
-        patomic_##opsig_kind##_store_t fp_store;                \
-        patomic_##opsig_kind##_load_t fp_load;                  \
-        patomic_##ops_kind##_xchg_t xchg_ops;                   \
-        patomic_##ops_kind##_bitwise_t bitwise_ops;             \
-        patomic_##ops_kind##_binary_t binary_ops;               \
-        patomic_##ops_kind##_arithmetic_t signed_ops;           \
-        patomic_##ops_kind##_arithmetic_t unsigned_ops;         \
-    } patomic_##ops_kind##_t
+    PATOMIC_DEFINE_XCHG_OPS_STRUCT(opsig_kind, ops_kind)
 
 /* implicit */
-PATOMIC_DEFINE_OPS_STRUCT(opsig, ops);
+PATOMIC_DEFINE_OPS_STRUCTS(opsig, ops);
+typedef struct {
+    patomic_opsig_store_t fp_store;
+    patomic_opsig_load_t fp_load;
+    patomic_ops_xchg_t xchg_ops;
+    patomic_ops_bitwise_t bitwise_ops;
+    patomic_ops_binary_t binary_ops;
+    patomic_ops_arithmetic_t signed_ops;
+    patomic_ops_arithmetic_t unsigned_ops;
+} patomic_ops_t;
 
 /* explicit */
-PATOMIC_DEFINE_OPS_STRUCT(opsig_explicit, ops_explicit);
+PATOMIC_DEFINE_OPS_STRUCTS(opsig_explicit, ops_explicit);
+typedef struct {
+    patomic_opsig_explicit_store_t fp_store;
+    patomic_opsig_explicit_load_t fp_load;
+    patomic_ops_explicit_xchg_t xchg_ops;
+    patomic_ops_explicit_bitwise_t bitwise_ops;
+    patomic_ops_explicit_binary_t binary_ops;
+    patomic_ops_explicit_arithmetic_t signed_ops;
+    patomic_ops_explicit_arithmetic_t unsigned_ops;
+} patomic_ops_explicit_t;
 
-#undef PATOMIC_DEFINE_OPS_STRUCT
+/* transaction */
+PATOMIC_DEFINE_OPS_STRUCTS(opsig_transaction, ops_transaction);
+typedef struct {
+    patomic_opsig_transaction_store_t fp_store;
+    patomic_opsig_transaction_load_t fp_load;
+    patomic_ops_transaction_xchg_t xchg_ops;
+    patomic_ops_transaction_bitwise_t bitwise_ops;
+    patomic_ops_transaction_binary_t binary_ops;
+    patomic_ops_transaction_arithmetic_t signed_ops;
+    patomic_ops_transaction_arithmetic_t unsigned_ops;
+    /* extra ops not available in implicit/explicit */
+    patomic_ops_transaction_flag_t flag_ops;
+    patomic_ops_transaction_special_t special_ops;
+} patomic_ops_transaction_t;
+
+
+#undef PATOMIC_DEFINE_OPS_STRUCTS
 #undef PATOMIC_DEFINE_XCHG_OPS_STRUCT
 #undef PATOMIC_DEFINE_BITWISE_OPS_STRUCT
 #undef PATOMIC_DEFINE_BINARY_OPS_STRUCT